// api/models/InterfaceInvokeTestRequest.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type InterfaceInvokeTestRequest = {
    id?: number;
    method?: string;
    name?: string;
    paramName?: string;
    requestParams?: string;
    url?: string;
};